feat: fixing the coding

+++ models/update.go
@@ -2,7 +2,7 @@ package models
 
 import "api-postgresql/db"
 
-func Update(id int, todo Todo) (int64, error) {
+func Update(id int64, todo Todo) (int64, error) {
 	conn, err := db.OpenConnection()
 	if err != nil {
 		return 0, err
